Use Cluster to fork

+++ main.js
@@ -5,8 +5,9 @@
  *
  */
 
-var logger = require('winston')
+var cluster = require('cluster')
   , fs = require('fs')
+  , logger = require('winston')
   , net = require('net')
   , os = require('os')
   , sugar = require('sugar')
@@ -14,40 +15,52 @@ var logger = require('winston')
 
 var Sentry = require('./sentry').Sentry;
 
-var SOCKET_FILE = os.tmpDir() + '/sentry-' + Date.now() + '.sock';
-
 function setupLogging(log) {
   log.remove(logger.transports.Console);
   log.add(logger.transports.Console, { colorize: true, timestamp: true });
 }
 
 function setupSignals() {
-  process.on('exit', cleanUp);
   process.on('SIGINT', function() {
     process.exit(1);
   });
-}
-
-function cleanUp() {
-  try {
-    fs.unlinkSync(SOCKET_FILE);
-
-  } catch (err) {
-    logger.warn(err);
-  }
-  logger.info('Exiting Sentry');
 }
 
 function main() {
   setupLogging(logger);
   setupSignals();
 
-  var sentry = new Sentry();
+  if (cluster.isMaster) {
+    for (var i = 0; i < os.cpus().length; i++) {
+      cluster.fork();
+    }
 
-  var server = net.createServer(function(c) {});
-  server.listen(SOCKET_FILE, function() {
-    logger.info('Server started (' + SOCKET_FILE + ').');
-  });
+    cluster.on('exit', function(worker, code, signal) {
+      logger.warn('Worker ' + worker.process.pid + ' died.');
+    });
+
+  } else {
+    setupLogging(logger);
+    setupSignals();
+
+    var sentry = new Sentry();
+    var SOCKET_FILE = os.tmpDir() + '/sentry-' + sentry.getId() + '.sock';
+
+    process.on('exit', function () {
+      try {
+        fs.unlinkSync(SOCKET_FILE);
+
+      } catch (err) {
+        logger.warn(err);
+      }
+      logger.info('Exiting Sentry');
+    });
+
+    var server = net.createServer(function(c) {});
+    server.listen(SOCKET_FILE, function() {
+      logger.info('Server started (' + SOCKET_FILE + ').');
+    });
+  }
 }
 
 main();
